fixed the error formatting (again)
This time with much more real-world testing of various scenarious such as shorter-than-expected backtraces (something y_testing makes impossible). This change removes the vscode friendly formatting unfortunately but it probably wasn't _that_ useful and log-parsers can do this anyway (like sampctl) faster, using regex etc.

The goal of this change is just to simplify the code so less can go wrong in the error handling code. It simply strips off the first two lines of the backtrace:

"AMX backtrace:""
"#0 native GetBacktrace () in crashdetect.DLL"

which are always the same regardless of the stack.

It doesn't attempt to re-format the stack trace lines, instead of just trims off the error number as it's offset by 2 anyway. This is done by simply finding the first space on the line. The address and all other information is exactly as it is in the original trace.

It would be nice if crashdetect provided this information in a nicer format for machine-readability but oh well.

Test updated accordingly.